+++ NEWS.md
@@ -1,4 +1,5 @@
-GitStats 0.1.0.9001
+
+GitStats 0.1.0.9002
 
 - set new name for `set_connection` function: `set_host` as it is more informative (and shorter) (I: #271),
 - changed name of a function: `add_team_member` to `set_team_member` (I: #271),
@@ -6,6 +7,7 @@ GitStats 0.1.0.9001
 - added `get_users()` function to pull information on users (I: #199 PR: #238),
 - added possibility of scanning whole internal git platforms if no `orgs` are passed (I: #258),
 - added switching to REST engine in case GraphQL fails with 502 error (I: #225 PRs: #227 (for repos) #261 (for commits))
+- added `show_orgs()` function to print all organizations (I: #283),
 - added GraphQL engine for getting GitLab repos by organization (I: #218 PR: #233)
 - removed `contributors` as basic stat when pulling `repos` by `org` and by `phrase` to improve speed of pulling repositories data. Added `get_repos_contributors()` user function and `add_contributors` parameter to `get_repos()` function to add conditionally information on contributors to repositories table (I: #235 PRs: #243 #264)
 - added resetting all settings to default with `reset()` function (I: #270)
@@ -13,7 +15,7 @@ GitStats 0.1.0.9001
 - OOP optimization: moved method on adding issues do repository table via REST to privates (I: #235 PR: #243)
 - handled errors when tokens do not grant access (I: #242 PR: #247)
 - in repositories output set `api_url` column as an address to the repo, not the host (I: #201 PR: #249)
-
+- fixed adding GitLab subgroups (I: #176)
 
 GitStats 0.1.0